+++ test/app/components/css_component.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class CssComponent < ViewComponent::Base
-  include ViewComponent::Stylable
+  include ViewComponent::Styleable
 
   def call
     content_tag(:div, "Hello, World!", class: styles["foo"])