+++ test/view_component/view_component_test.rb
@@ -709,7 +709,7 @@ class ViewComponentTest < ViewComponent::TestCase
     assert_selector("style", visible: false, count: 1)
   end
 
-  def test_renders_stylable_component_selector
+  def test_renders_styleable_component_selector
     render_inline(CssComponent.new)
 
     assert_selector(".#{CssComponent.styles['foo']}")